Updates and cleanup from review, simplify auto cutoff calculation Renamed variables to be more representative of their content.
Simplified the auto-cutoff calculation to be 90% of the nyquist frequency of the measured rx frame rate.

Simplified the PT1 from BIQUAD calculation.

Added active cutoffs to the blackbox log header. Reduce the number of headers by combining like entries into a single line.

+++ src/main/fc/fc_rc.h
@@ -39,4 +39,4 @@ void updateRcCommands(void);
 void resetYawAxis(void);
 void initRcProcessing(void);
 bool isMotorsReversed(void);
-uint16_t rcSmoothingGetValue(int whichValue);
+int rcSmoothingGetValue(int whichValue);
